#1 トップページから編集削除リンクを取り除いて、manageページを追加したよ

+++ database/migrations/2019_10_02_045401_create_comments_table.php
@@ -16,7 +16,8 @@ class CreateCommentsTable extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->bigIncrements('id');
             // $table->integer('post_id');
-            $table->unsignedInteger('post_id');
+            $table->unsignedBigInteger('post_id');
+            // $table->unsignedInteger('post_id');
             $table->string('body');
             $table->timestamps();
             $table
